refactor: centralize supabase client and add status indexes
