+++ src/modules/usuarios/auth/index.ts
@@ -1,2 +1,2 @@
-export { supabase } from "./supabase-client";
+export { supabase } from "../../../config/supabase";
 export { supabaseAuthMiddleware } from "./supabase-middleware";
